refactor: rm useless hooks
