+++ web-registry/src/components/organisms/ProjectTopSection/ProjectTopSection.utils.ts
@@ -1,6 +1,19 @@
 import { User } from 'web-components/lib/components/user/UserInfo';
 import { truncate } from 'web-components/lib/utils/truncate';
 
+import {
+  CreditClass,
+  CreditClassVersion,
+  Maybe,
+  Project,
+} from 'generated/graphql';
+import {
+  AnchoredProjectMetadataIntersectionLD,
+  ProjectPageMetadataLD,
+  ProjectQuote,
+} from 'lib/db/types/json-ld';
+import { getAreaUnit, qudtUnit } from 'lib/rdf';
+
 // TODO
 // This is a temporary hack to show Regen as a Project Admin when applicable
 
@@ -25,5 +38,89 @@ export const getDisplayAdmin = (address?: string): User | undefined => {
     name: truncate(address),
     type: 'ORGANIZATION',
     link: `/ecocredits/accounts/${address}`,
+  };
+};
+
+type AnchoredMetadata = {
+  projectName?: string;
+  area?: string;
+  areaUnit?: string;
+  placeName?: string;
+};
+
+export const getAnchoredMetadata = (
+  anchoredMetadata?: Partial<AnchoredProjectMetadataIntersectionLD>,
+): AnchoredMetadata => {
+  const projectName = anchoredMetadata?.['schema:name'];
+  const projectSize = anchoredMetadata?.['regen:projectSize'];
+  const area = projectSize?.['qudt:numericValue']?.['@value'];
+  const unit = projectSize?.['qudt:unit']?.['@value'];
+  const areaUnit = getAreaUnit(unit as qudtUnit);
+  const placeName =
+    anchoredMetadata?.['schema:location']?.['place_name'] ??
+    anchoredMetadata?.['schema:location']?.['geojson:place_name'];
+
+  return { projectName, area, areaUnit, placeName };
+};
+
+type UnanchoredMetadata = {
+  videoURL?: string | null;
+  glanceText?: string[];
+  primaryDescription?: string;
+  quote?: ProjectQuote;
+  landStewardPhoto?: string | null;
+  landStewardStoryTitle?: string | null;
+  landStewardStory?: string | null;
+};
+
+export const getUnanchoredMetadata = (
+  projectPageMetadata?: Partial<ProjectPageMetadataLD>,
+): UnanchoredMetadata => {
+  const videoURL = projectPageMetadata?.['regen:videoURL']?.['@value'];
+  const glanceText = projectPageMetadata?.['regen:glanceText']?.['@list'];
+  const primaryDescription =
+    projectPageMetadata?.['regen:landStory'] ||
+    projectPageMetadata?.['schema:description'];
+  const quote = projectPageMetadata?.['regen:projectQuote'];
+  const landStewardPhoto =
+    projectPageMetadata?.['regen:landStewardPhoto']?.['@value'];
+  const landStewardStoryTitle =
+    projectPageMetadata?.['regen:landStewardStoryTitle'];
+  const landStewardStory = projectPageMetadata?.['regen:landStewardStory'];
+
+  return {
+    videoURL,
+    glanceText,
+    primaryDescription,
+    quote,
+    landStewardPhoto,
+    landStewardStoryTitle,
+    landStewardStory,
+  };
+};
+
+type ProjectDetails = {
+  creditClass?: Maybe<CreditClass>;
+  creditClassVersion?: Maybe<CreditClassVersion>;
+  sdgIris?: any[];
+  offsetGenerationMethod?: string;
+};
+
+export const getProjectDetails = (project?: Maybe<Project>): ProjectDetails => {
+  const creditClass = project?.creditClassByCreditClassId;
+  const creditClassVersion = creditClass?.creditClassVersionsById?.nodes?.[0];
+  const sdgIris = creditClassVersion?.metadata?.['http://regen.network/SDGs']?.[
+    '@list'
+  ]?.map((sdg: { '@id': string }) => sdg['@id']);
+  const offsetGenerationMethod =
+    creditClassVersion?.metadata?.[
+      'http://regen.network/offsetGenerationMethod'
+    ];
+
+  return {
+    creditClass,
+    creditClassVersion,
+    sdgIris,
+    offsetGenerationMethod,
   };
 };
